\name{cooptrees-package}
\alias{cooptrees-package}
\alias{cooptrees}
\docType{package}
\title{Cooperative aspects of optimal trees in weighted graphs}
\description{Computes several cooperative games and allocation rules associated with minimum cost spanning tree problems and minimum cost arborescence problems.}
\details{
\tabular{ll}{
Package: \tab cooptrees\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-01\cr
License: \tab GPL-3\cr
}
The most important functions are \link{mstCooperative} and \link{maCooperative}. The other functions included in the package are auxiliary ones that can be used independently.
}
\author{
Manuel Fontenla

Maintainer: Manuel Fontenla <manu.fontenla@gmail.com>
}
\references{
B. Dutta and D. Mishra, "Minimum cost arborescences", Games
and Economic Behavior, vol. 74, pp. 120-143, Jan. 2012.

C. G. Bird, "On Cost Allocation for a Spanning Tree: A Game
Theoretic Approach", Networks, no. 6, pp. 335-350, 1976.

B. Dutta and A. Kar, "Cost monotonicity, consistency and
minimum cost spanning tree games", Games and Economic
Behavior, vol. 48, pp. 223-248, Aug. 2004.

A. Kar, "Axiomatization of the Shapley Value on Minimum
Cost Spanning Tree Games", Games and Economic Behavior,
vol. 38, pp. 265-277, Feb. 2002.

G. Bergantiños and J. J. Vidal-Puga, "A fair rule in
minimum cost spanning tree problems", Journal of Economic
Theory, vol. 137, pp. 326-352, Nov. 2007.

G. Bergantiños and J. J. Vidal-Puga, "The optimistic TU
game in minimum cost spanning tree problems", International
Journal of Game Theory, vol. 36, pp. 223-239, Feb. 2007.

V. Feltkamp, S. H. Tijs, S. Muto, "On the irreducible core
and the equal remaining obligation rule of minimum cost
extension problems", Mimeo, Tilburg University, 1994.
}
\encoding{UTF-8}